agents/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
av==14.4.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-genai==1.22.0
googleapis-common-protos==1.70.0
gradio==5.34.2
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
livekit==1.0.11
livekit-agents==1.1.4
livekit-api==1.0.3
livekit-plugins-google==1.1.4
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.1.4
livekit-protocol==1.0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.1
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.92.2
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruff==0.12.1
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.46.2
tenacity==8.5.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
-e git+https://github.com/livekit/agents.git@74fdb91c2f81c7d75e5b48acb53bb54562d7cb5f#egg=UNKNOWN
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
